Better edge-case handling when braiding rows: ensure empty column sets do not erase result rows
Before, if a column set had no rows, when braiding it with others, the result would always be 0 rows. It is basically like multiplying the column sets. Thats not what we want, so now we detect empty column sets and just eliminate them, continuing to run the algorithm on the remaining column sets. Also added an early out for the now-possible case where there is exactly one column set we are trying to braid - that makes no sense, the column set should just be the result.